requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
APScheduler==3.6.3
async-timeout==3.0.1
attrs==21.2.0
Backtesting==0.3.1
backtrader==1.9.76.123
binance==0.3
bokeh==2.3.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
dateparser==1.0.0
decorator==5.0.9
ffn==0.3.6
future==0.18.2
idna==2.10
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mpl-finance==0.10.1
mplfinance==0.12.7a17
multidict==5.1.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
pefile==2021.5.24
Pillow==8.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyparsing==2.4.7
PyPrind==2.11.3
python-binance==1.0.12
python-dateutil==2.8.1
python-telegram-bot==13.6
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
ta==0.7.0
tabulate==0.8.9
threadpoolctl==2.1.0
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.5
websockets==9.1
yarl==1.6.3
